requirements.txt
# Python >= 3.8 required
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.3.0,<2.0.0
pydantic>=2.4.2,<3.0.0
